Move some files into corresponding subdirectory
Fix some bugs where the MainController is still null when trying
to reference from it. Fix a potential bug with processAuthRequest where
the authorized boolean will still be undefined when reaching
refreshTokens().
